# Conflicts:
#	Kvazzup.pro
#	src/kvazzupcontroller.cpp
#	src/media/delivery/delivery.cpp
#	src/ui/gui/callwindow.cpp
#	src/ui/gui/conferenceview.cpp
#	src/ui/settings/defaultsettings.h
#	src/ui/uimanager.cpp
#	src/ui/uimanager.h
#	src/videoviewfactory.cpp

+++ src/icetypes.h
@@ -40,6 +40,6 @@ struct ICEPair
 {
   std::shared_ptr<ICEInfo> local;
   std::shared_ptr<ICEInfo> remote;
-  int priority;
+  uint64_t priority;
   PairState state;
 };
